Making sure volume is between 0 & 1

+++ commands/setvolume.js
@@ -6,7 +6,14 @@ module.exports = {
   numArgs:2,
   usage: '<soundcommand> <volume>',
   execute(message, args,client) {
-    client.getDbHelper().addCommandVolume(args[0],args[1]);
-    message.delete().catch(err => console.log(err));
+    var volume = args[1];
+    if((volume-1)*(volume-0) <= 0){
+      client.getDbHelper().addCommandVolume(args[0],args[1]);
+      message.delete().catch(err => console.log(err));
+    }
+    else{
+        var reply = 'Volume must be between 0 and 1';
+        message.reply(reply);
+    }
   },
 };
